+++ md2cf/document.py
@@ -7,6 +7,7 @@ import mistune
 import yaml
 from yaml.parser import ParserError
 
+from md2cf.ignored_files import GitRepository
 from md2cf.confluence_renderer import ConfluenceRenderer
 
 
@@ -21,6 +22,7 @@ class Page(object):
         parent_id: str = None,
         parent_title: str = None,
         space: str = None,
+        labels: Optional[List[str]] = None,
     ):
         self.title = title
         self.body = body
@@ -32,17 +34,36 @@ class Page(object):
         self.parent_id = parent_id
         self.parent_title = parent_title
         self.space = space
+        self.labels = labels
 
     def get_content_hash(self):
         return hashlib.sha1(self.body.encode()).hexdigest()
 
+    def __repr__(self):
+        return "Page({})".format(
+            ", ".join(
+                [
+                    "{}={}".format(name, repr(value))
+                    for name, value in [
+                        ["title", self.title],
+                        ["file_path", self.file_path],
+                        ["page_id", self.page_id],
+                        ["parent_id", self.parent_id],
+                        ["parent_title", self.parent_title],
+                        ["space", self.space],
+                    ]
+                ]
+            )
+        )
+
 
 def find_non_empty_parent_path(
-    current_dir: Path, folder_data: Dict[Path, Dict[str, Any]]
+    current_dir: Path, folder_data: Dict[Path, Dict[str, Any]], default: Path
 ) -> Path:
     for parent in current_dir.parents:
-        if folder_data[parent]["n_files"]:
+        if parent in folder_data and folder_data[parent]["n_files"]:
             return parent
+    return default.absolute()
 
 
 def get_pages_from_directory(
@@ -53,11 +74,29 @@ def get_pages_from_directory(
     beautify_folders: bool = False,
     use_pages_file: bool = False,
     strip_header: bool = False,
+    remove_text_newlines: bool = False,
+    use_gitignore: bool = True,
 ) -> List[Page]:
+    """
+    Collect a list of markdown files recursively under the file_path directory.
+
+    :param file_path: The starting path from which to search
+    :param collapse_single_pages:
+    :param skip_empty:
+    :param collapse_empty:
+    :param beautify_folders:
+    :param use_pages_file:
+    :param strip_header:
+    :param remove_text_newlines:
+    :param use_gitignore: Use .gitignore files to skip unwanted markdown in directory search
+    :return: A list of paths to the markdown files to upload.
+    """
     processed_pages = list()
     base_path = file_path.resolve()
     parent_page_title = None
     folder_data = dict()
+    git_repo = GitRepository(file_path, use_gitignore=use_gitignore)
+
     for current_path, directories, file_names in os.walk(file_path):
         current_path = Path(current_path).resolve()
 
@@ -65,6 +104,10 @@ def get_pages_from_directory(
             Path(current_path, file_name)
             for file_name in file_names
             if file_name.endswith(".md")
+        ]
+        # Filter out ignored files
+        markdown_files = [
+            path for path in markdown_files if not git_repo.is_ignored(path)
         ]
 
         folder_data[current_path] = {
@@ -82,19 +125,18 @@ def get_pages_from_directory(
             # TODO: add support for .pages file to read folder title
             if skip_empty or collapse_empty:
                 folder_parent_path = find_non_empty_parent_path(
-                    current_path, folder_data
+                    current_path, folder_data, default=file_path
                 )
             else:
                 folder_parent_path = current_path.parent
 
             folder_parent_title = folder_data[folder_parent_path]["title"]
-
             if len(markdown_files) == 1 and collapse_single_pages:
                 parent_page_title = folder_parent_title
             else:
                 if collapse_empty:
-                    folder_data[current_path]["title"] = current_path.relative_to(
-                        folder_parent_path
+                    folder_data[current_path]["title"] = str(
+                        current_path.relative_to(folder_parent_path)
                     )
                 if beautify_folders:
                     folder_data[current_path]["title"] = (
@@ -121,7 +163,9 @@ def get_pages_from_directory(
 
         for markdown_file in markdown_files:
             processed_page = get_page_data_from_file_path(
-                markdown_file, strip_header=strip_header
+                markdown_file,
+                strip_header=strip_header,
+                remove_text_newlines=remove_text_newlines,
             )
             processed_page.parent_title = parent_page_title
             processed_pages.append(processed_page)
@@ -135,14 +179,20 @@ def get_pages_from_directory(
     return processed_pages
 
 
-def get_page_data_from_file_path(file_path: Path, strip_header: bool = False) -> Page:
+def get_page_data_from_file_path(
+    file_path: Path, strip_header: bool = False, remove_text_newlines: bool = False
+) -> Page:
     if not isinstance(file_path, Path):
         file_path = Path(file_path)
 
     with open(file_path) as file_handle:
         markdown_lines = file_handle.readlines()
 
-    page = get_page_data_from_lines(markdown_lines, strip_header=strip_header)
+    page = get_page_data_from_lines(
+        markdown_lines,
+        strip_header=strip_header,
+        remove_text_newlines=remove_text_newlines,
+    )
 
     if not page.title:
         page.title = file_path.stem
@@ -153,21 +203,43 @@ def get_page_data_from_file_path(file_path: Path, strip_header: bool = False) ->
 
 
 def get_page_data_from_lines(
-    markdown_lines: List[str], strip_header: bool = False
+    markdown_lines: List[str],
+    strip_header: bool = False,
+    remove_text_newlines: bool = False,
 ) -> Page:
     frontmatter = get_document_frontmatter(markdown_lines)
     if "frontmatter_end_line" in frontmatter:
         markdown_lines = markdown_lines[frontmatter["frontmatter_end_line"] :]
 
-    page = parse_page(markdown_lines, strip_header=strip_header)
+    page = parse_page(
+        markdown_lines,
+        strip_header=strip_header,
+        remove_text_newlines=remove_text_newlines,
+    )
 
     if "title" in frontmatter:
         page.title = frontmatter["title"]
+
+    if "labels" in frontmatter:
+        if isinstance(frontmatter["labels"], list):
+            page.labels = [str(label) for label in frontmatter["labels"]]
+        else:
+            raise TypeError(
+                "the labels section in the frontmatter " "must be a list of strings"
+            )
     return page
 
 
-def parse_page(markdown_lines: List[str], strip_header: bool = False) -> Page:
-    renderer = ConfluenceRenderer(use_xhtml=True, strip_header=strip_header)
+def parse_page(
+    markdown_lines: List[str],
+    strip_header: bool = False,
+    remove_text_newlines: bool = False,
+) -> Page:
+    renderer = ConfluenceRenderer(
+        use_xhtml=True,
+        strip_header=strip_header,
+        remove_text_newlines=remove_text_newlines,
+    )
     confluence_mistune = mistune.Markdown(renderer=renderer)
     confluence_content = confluence_mistune("".join(markdown_lines))