+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Support for labels in the Markdown front matter
+- Option to replace all labels with the set specified in the Markdown front matter
+- Option to add a postface to every page
+- Option to ignore `.gitignore` files (@bjorns)
+- Option to remove single newlines from paragraphs (@bjorns)
+### Changed
+- Skip files and directories listed in `.gitignore` files (@bjorns)
+
 ## 1.4.0 - 2022-07-27
 ### Added
 - Add option to remove the top level header from the page